% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataacq.R
\docType{data}
\name{alcons}
\alias{alcons}
\title{alcons function}
\format{An object of class \code{data.frame} with 193 rows and 5 columns.}
\usage{
alcons
}
\description{
The "alcons" function is an easy way to visualize the dataset

The function needs the argument "alcons".
The output shows a graph built using the "ggplot2 package"
}
\author{
Anna Errichiello <anna.errichiello@studenti.unimi.it>
}
\keyword{datasets}
