% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_leverage.R
\name{pca_leverage}
\alias{pca_leverage}
\title{Calculate the leverage for a Principal Component Analysis}
\usage{
pca_leverage(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculate the leverage for a Principal Component Analysis
}
