% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotalcons.R
\name{plotalcons}
\alias{plotalcons}
\title{plotalcons function}
\usage{
plotalcons()
}
\description{
The "plotalcons" function is an easy way to visualize the dependence of "beer servings" on the considered country.

The output shows a graph built using the "ggplot2 package".
}
\author{
Anna Errichiello <anna.errichiello@studenti.unimi.it>
}
